@@ -3,7 +3,7 @@ package com.example.sumon.androidvolley.utils;
 import com.android.volley.toolbox.ImageLoader.ImageCache;
 
 import android.graphics.Bitmap;
-import android.support.v4.util.LruCache;
+import androidx.collection.LruCache;
 
 public class LruBitmapCache extends LruCache<String, Bitmap> implements ImageCache {
 	public static int getDefaultLruCacheSize() {
